flyttat interface och repository till rätt map och gjort crud med ef
