+++ HenriksHobbyLager/Interface/IRepository.cs
@@ -8,10 +8,10 @@ namespace HenriksHobbylager.Repositories;
 
 public interface IRepository<T> where T : class
 {
-	T GetById(int id);
+	Task<IEnumerable<T>> GetAllAsync(Expression<Func<T, bool>> predicate);
 	void CreateProduct(T entity);
 	void Update(T entity);
 	void Delete(int id);
-	IEnumerable<T> GetAll();
 	IEnumerable<T> Search(Func<T, bool> predicate);
+	Task SaveChangesAsync();
 }
